+++ src/all/background_page/service/api/resource/resourceService.js
@@ -195,6 +195,37 @@ class ResourceService extends AbstractService {
     const response = await this.apiClient.fetchAndHandleResponse('GET', url);
     return response.body;
   }
+
+  /**
+   * Find resources to export
+   * @param {array} resourcesIds
+   * @returns {Promise<*>} Response body
+   * @public
+   */
+  async findAllForExport(resourcesIds) {
+    // Retrieve by batch to avoid any 414 response.
+    const batchSize = 80;
+    if (resourcesIds.length > batchSize) {
+      let resources = [];
+      const totalBatches = Math.ceil(resourcesIds.length / batchSize);
+      for (let i = 0; i < totalBatches; i++) {
+        const resourcesIdsPart = resourcesIds.splice(0, batchSize);
+        const resourcesPart = await this.findAllForExport(resourcesIdsPart);
+        resources = [...resources, ...resourcesPart];
+      }
+
+      return resources;
+    }
+
+    let url = this.apiClient.buildUrl(this.apiClient.baseUrl.toString());
+    resourcesIds.forEach(resourceId => {
+      url.searchParams.append(`filter[has-id][]`, resourceId);
+    });
+    url.searchParams.append('contain[secret]', '1');
+
+    const response = await this.apiClient.fetchAndHandleResponse('GET', url);
+    return response.body;
+  }
 }
 
 exports.ResourceService = ResourceService;